+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2026-08-10 (In Progress)
+
+### 🚀 New Features
+
+#### File Transfer Enhancement
+- [x] Remove file size limit
+- [x] Show file in chat directly
+- [x] Transfer progress bar
+- [x] Transfer speed display
+
+#### Connection Quality Info
+- [x] Display peer IP and port
+- [x] Show latency/ping
+- [x] Connection quality indicator
+
+---
+
 ## [1.0.5] - 2026-08-10
 
 ### 🐛 Bug Fixes (3 fixes)